Implemented Payment Screen And New Local User Table
Includes Credit Card Entry and Option for Paypal. Entry is yet to be
implemented and validated into the remote database.
Paypal not implemented yet.
Also implemented insertion of user data to local database. This info is
needed for queries later on. Such as when we need the users unique email
address.